scripts/containers/cwl-ping/pyproject.toml
[project]
name = "cwl-ping"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.13.2",
    "asyncpg>=0.31.0",
    "python-dotenv>=1.2.1",
]
